+++ src/Smart.Design.Razor/TagHelpers/ValidationMessage/TagNames.cs
@@ -1,5 +1,12 @@
 namespace Smart.Design.Razor.TagHelpers.ValidationMessage;
+
+/// <summary>
+/// Defines the Tag name related to the Validation Messages components.
+/// </summary>
 public static class TagNames
 {
+    /// <summary>
+    /// Tag name of <see cref="ValidationMessageTagHelper"/>.
+    /// </summary>
     public const string ValidationMessageTagName = "smart-validation-message";
 }
